requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
lxml==4.9.3
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.7